# html-starter-kit
# html-starter-kit
##Overview
This is HTML quick start pack for basic level web developer.For working
in local host or your local computer without internet conncetion, you
can use it.
Every developer should not start with minified code because it hides
your eyes from looking the most important codes.
As you are going to be a  the coder and write  down in your heart, try
to understand the codes.
## Supported  Features
1.Supported with Bootstrap Responsive Framework
2.Supported with Font Awesome
3.Used in Customized Font Folder
4.Responsive Framework
5.No-Minified Code.